cnc fixx

+++ scripts/cnc.py
@@ -294,13 +294,14 @@ def normalize_dump(service, input_files, output_dir, newdump = False):
 
 
     # remove duplicates from entities_with_no_omid
-    index_entities = set()
-    with open(output_dir+'entities_with_no_omid.csv') as csv_file:
-        csv_reader = csv.reader(csv_file, delimiter=',')
-        for row in csv_reader:
-            index_entities.add(row[0])
-    with open(output_dir+'entities_with_no_omid.csv', 'w') as f_out:
-        csv.writer(f_out).writerows([[e] for e in index_entities])
+    if os.path.exists(output_dir+'entities_with_no_omid.csv'):
+        index_entities = set()
+        with open(output_dir+'entities_with_no_omid.csv') as csv_file:
+            csv_reader = csv.reader(csv_file, delimiter=',')
+            for row in csv_reader:
+                index_entities.add(row[0])
+        with open(output_dir+'entities_with_no_omid.csv', 'w') as f_out:
+            csv.writer(f_out).writerows([[e] for e in index_entities])
 
 def main():
     global _config
